Reorganize user components.
