+++ client/src/app/modules/auth/modules/employee/components/users/user-detail/user-detail.component.html
